pyproject.toml
[project]
name = "health-livekit-insurance-agent"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "livekit-agents[deepgram,elevenlabs,openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv~=1.0.0",
    "pydantic~=2.0",
    "httpx~=0.27.0",
    "structlog~=24.0",
    "tenacity~=8.0",
]

[project.optional-dependencies]
dev = ["pytest~=8.0", "pytest-asyncio~=0.23", "black~=24.0", "ruff~=0.3"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.26.0",
]
